# smv2/openapi.json
{
  "components": {
    "schemas": {
...
      "ChapterOut": {
        "properties": {
          "answers_section_ids": {
            "items": {
              "type": "string"
            },
            "title": "Answers Section Ids",
            "type": "array"
          },
          "chapter_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Label"
          },
          "practice_section_ids": {
            "items": {
              "type": "string"
            },
            "title": "Practice Section Ids",
            "type": "array"
          },
          "section_ids": {
            "items": {
              "type": "string"
            },
            "title": "Section Ids",
            "type": "array"
          },
          "test_stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChapterTestStatsOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "chapter_label",
          "section_ids",
          "practice_section_ids",
          "answers_section_ids",
          "test_stats"
        ],
        "title": "ChapterOut",
        "type": "object"
      },
      "ChapterTestStatsOut": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "best_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Score"
          },
          "latest_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Score"
          }
        },
        "required": [
          "attempts",
          "best_score",
          "latest_score"
        ],
        "title": "ChapterTestStatsOut",
        "type": "object"
      },
      "ChatCitationOut": {
        "properties": {
          "n": {
            "title": "N",
            "type": "integer"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "section_id": {
            "title": "Section Id",
            "type": "string"
          },
          "source_ref": {
            "title": "Source Ref",
            "type": "string"
          }
        },
        "required": [
          "n",
          "section_id",
          "page",
          "source_ref"
        ],
        "title": "ChatCitationOut",
        "type": "object"
      },
      "ChatIn": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatSelectionIn"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatIn",
        "type": "object"
      },
      "ChatOut": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/ChatCitationOut"
            },
            "title": "Citations",
            "type": "array"
          },
          "reply_md": {
            "title": "Reply Md",
            "type": "string"
          }
        },
        "required": [
          "reply_md",
          "citations"
        ],
        "title": "ChatOut",
        "type": "object"
      },
      "ChatSelectionIn": {
        "description": "A passage the student selected in the reader \u2014 same 2000-char cap as\nHighlightIn.exact. Grounds this turn in that passage (ADR-024 feature).",
        "properties": {
          "exact": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Exact",
            "type": "string"
          },
          "section_id": {
            "title": "Section Id",
            "type": "string"
          }
        },
        "required": [
          "section_id",
          "exact"
        ],
        "title": "ChatSelectionIn",
        "type": "object"
      },
      "ChatTurnOut": {
        "properties": {
          "citations": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citations"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "citations",
          "created_at"
        ],
        "title": "ChatTurnOut",
        "type": "object"
      },
...
      "CourseOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failed_asset_count": {
            "default": 0,
            "title": "Failed Asset Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgressSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "section_count": {
            "default": 0,
            "title": "Section Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "CourseOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProgressSummary": {
        "properties": {
          "scroll_pos": {
            "title": "Scroll Pos",
            "type": "number"
          },
          "section_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "section_id",
          "scroll_pos",
          "updated_at"
        ],
        "title": "ProgressSummary",
        "type": "object"
      },
...
      "StudyNextItemOut": {
        "description": "One deterministic study suggestion (ADR-022, app.services.study_service).\n`detail` holds whatever raw numbers back `reason` \u2014 {\"best_score\":\n...} for low_test_score, {\"due_count\": ...} for due_cards, {} for\nunread, {\"days_since\": ...} for stale.",
        "properties": {
          "chapter_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Label"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "reason": {
            "enum": [
              "low_test_score",
              "due_cards",
              "unread",
              "stale"
            ],
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "chapter_label",
          "reason",
          "detail"
        ],
        "title": "StudyNextItemOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SourceMind v2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/courses/{course_id}": {
      "delete": {
        "operationId": "delete_course",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Course",
        "tags": [
          "courses"
        ]
      },
      "get": {
        "operationId": "get_course",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Course",
        "tags": [
          "courses"
        ]
      }
    },
...
    "/api/courses/{course_id}/chapters": {
      "get": {
        "operationId": "list_chapters",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChapterOut"
                  },
                  "title": "Response List Chapters",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Chapters",
        "tags": [
          "chapters"
        ]
      }
    },
    "/api/courses/{course_id}/chat": {
      "get": {
        "operationId": "chat_history",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatTurnOut"
                  },
                  "title": "Response Chat History",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat History",
        "tags": [
          "chat"
        ]
      },
      "post": {
        "operationId": "send_chat",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Chat",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/api/courses/{course_id}/study-next": {
      "get": {
        "operationId": "study_next",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudyNextItemOut"
                  },
                  "title": "Response Study Next",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Study Next",
        "tags": [
          "chapters"
        ]
      }
    },
...
  }
}